requirements.txt
PyYAML
future
construct
configparser
google
google-api-core
google-api-python-client
google-cloud
google-cloud-bigquery
google-cloud-core
protobuf
logger
pip
protobuf
pycparser
pyparsing
pyzmq
packaging
python-dateutil
typing-extensions
dotmap


